Split pure helpers out of component modules
preloadZapLogo/drawPreloadedZapLogo move from images.tsx into lib/zapLogo.ts,
and isWriteModalOpen moves from WriteModal.tsx into writeNfcHook.tsx. Both were
pure functions living in modules that also export React components, so route
files that only wanted the helper pulled the component graph in with it.

Tests that mocked writeNfcHook wholesale now spread importOriginal() so the
enums and isWriteModalOpen stay real.
