+++ src/__tests__/unit/routes/create.custom.test.tsx
@@ -59,18 +59,10 @@ vi.mock("@/lib/preferencesStore", () => ({
   }),
 }));
 
-// Mock NFC writer
-vi.mock("@/lib/writeNfcHook", () => ({
+// Mock NFC writer, keeping the real enums and isWriteModalOpen
+vi.mock("@/lib/writeNfcHook", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("@/lib/writeNfcHook")>()),
   useNfcWriter: () => mockNfcWriter,
-  WriteMethod: {
-    Auto: "auto",
-    LocalNFC: "local",
-    RemoteReader: "remote",
-  },
-  WriteAction: {
-    Write: "write",
-    Read: "read",
-  },
 }));
 
 // Mock hooks
@@ -102,10 +94,6 @@ vi.mock("@/components/ZapScriptInput.tsx", () => ({
 
 // Mock WriteModal to simplify testing
 vi.mock("@/components/WriteModal", () => ({
-  isWriteModalOpen: (
-    writeIntent: boolean,
-    writer: { status: unknown; verifyError: unknown },
-  ) => writeIntent && (writer.status === null || writer.verifyError !== null),
   WriteModal: ({ isOpen }: { isOpen: boolean }) =>
     isOpen ? <div data-testid="write-modal">Write Modal</div> : null,
 }));
