+++ src/components/WriteModal.tsx
@@ -4,17 +4,9 @@ import { useSmartSwipe } from "@/hooks/useSmartSwipe";
 import { useBackButtonHandler } from "@/hooks/useBackButtonHandler";
 import { useFocusTrap } from "@/hooks/useFocusTrap";
 import { ScanResult } from "@/lib/models";
-import type { WriteNfcHook } from "@/lib/writeNfcHook";
 import { ScanSpinner } from "./ScanSpinner";
 import { Button } from "./wui/Button";
 import { useAnnouncer } from "./A11yAnnouncer";
-
-export function isWriteModalOpen(
-  writeIntent: boolean,
-  writer: Pick<WriteNfcHook, "status" | "verifyError">,
-): boolean {
-  return writeIntent && (writer.status === null || writer.verifyError !== null);
-}
 
 export function WriteModal(props: {
   isOpen: boolean;
