+++ src/routes/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { preloadZapLogo } from "@/lib/images";
+import { preloadZapLogo } from "@/lib/zapLogo";
 import { Index } from "./-pages/Index";
 
 export const Route = createFileRoute("/")({
